ASR-API/requirements.txt
websocket-client==1.6.4
urllib3==1.26.18
pyaudio==0.2.11
numpy==1.24.3 